move panel creation code around to consolidate
