+++ apps/ltn/src/design_ltn.rs
@@ -3,15 +3,15 @@ use street_network::Direction;
 use widgetry::mapspace::{DummyID, World};
 use widgetry::tools::{ChooseSomething, PopupMsg};
 use widgetry::{
-    Color, ControlState, DrawBaselayer, Drawable, EventCtx, GeomBatch, GfxCtx, Key, Line, Outcome,
-    Panel, State, TextExt, Toggle, Widget,
+    lctrl, Color, ControlState, DrawBaselayer, Drawable, EventCtx, GeomBatch, GfxCtx, Key, Line,
+    Outcome, Panel, RewriteColor, State, TextExt, Toggle, Widget,
 };
 
-use crate::components::{AppwidePanel, Mode};
+use crate::components::{AppwidePanel, BottomPanel, Mode};
 use crate::draw_cells::RenderCells;
 use crate::edit::{EditMode, EditNeighbourhood, EditOutcome};
 use crate::filters::auto::Heuristic;
-use crate::{colors, is_private, App, Neighbourhood, NeighbourhoodID, Transition};
+use crate::{colors, is_private, App, FilterType, Neighbourhood, NeighbourhoodID, Transition};
 
 pub struct DesignLTN {
     appwide_panel: AppwidePanel,
@@ -92,7 +92,7 @@ impl DesignLTN {
         };
         self.show_error = ctx.upload(show_error);
 
-        self.bottom_panel = self.edit.make_panel(
+        self.bottom_panel = make_bottom_panel(
             ctx,
             app,
             &self.appwide_panel,
@@ -391,4 +391,114 @@ fn advanced_panel(ctx: &EventCtx, app: &App) -> Widget {
             .build_def(ctx),
     ])
     .section(ctx)
+}
+
+fn make_bottom_panel(
+    ctx: &mut EventCtx,
+    app: &App,
+    appwide_panel: &AppwidePanel,
+    per_tab_contents: Widget,
+) -> Panel {
+    let row = Widget::row(vec![
+        if app.per_map.consultation.is_none() {
+            ctx.style()
+                .btn_outline
+                .text("Adjust boundary")
+                .hotkey(Key::B)
+                .build_def(ctx)
+                .centered_vert()
+            // TODO Vertical sep
+        } else {
+            Widget::nothing()
+        },
+        edit_mode(ctx, app),
+        match app.session.edit_mode {
+            EditMode::Filters => crate::edit::filters::widget(ctx),
+            EditMode::FreehandFilters(_) => crate::edit::freehand_filters::widget(ctx),
+            EditMode::Oneways => crate::edit::one_ways::widget(ctx),
+            EditMode::Shortcuts(ref focus) => {
+                crate::edit::shortcuts::widget(ctx, app, focus.as_ref())
+            }
+        }
+        .named("edit mode contents"),
+        ctx.style()
+            .btn_plain
+            .icon("system/assets/tools/undo.svg")
+            .disabled(app.per_map.edits.previous_version.is_none())
+            .hotkey(lctrl(Key::Z))
+            .build_widget(ctx, "undo"),
+        Widget::col(vec![
+            // TODO Only count new filters, not existing
+            format!(
+                "{} filters added",
+                app.per_map.edits.roads.len() + app.per_map.edits.intersections.len()
+            )
+            .text_widget(ctx),
+            format!(
+                "{} road directions changed",
+                app.per_map.edits.one_ways.len()
+            )
+            .text_widget(ctx),
+        ]),
+        per_tab_contents,
+    ]);
+
+    BottomPanel::new(ctx, appwide_panel, row)
+}
+
+fn edit_mode(ctx: &mut EventCtx, app: &App) -> Widget {
+    let edit_mode = &app.session.edit_mode;
+    let filter = |ft: FilterType, hide_color: Color, name: &str| {
+        let mut btn = ctx
+            .style()
+            .btn_solid_primary
+            .icon(ft.svg_path())
+            .image_color(
+                RewriteColor::Change(hide_color, Color::CLEAR),
+                ControlState::Default,
+            )
+            .image_color(
+                RewriteColor::Change(hide_color, Color::CLEAR),
+                ControlState::Disabled,
+            )
+            .disabled(matches!(edit_mode, EditMode::Filters) && app.session.filter_type == ft);
+        if app.session.filter_type == ft {
+            btn = btn.hotkey(Key::F1);
+        }
+        btn.build_widget(ctx, name)
+    };
+
+    Widget::row(vec![
+        Widget::row(vec![
+            filter(
+                FilterType::WalkCycleOnly,
+                Color::hex("#0b793a"),
+                "Modal filter -- walking/cycling only",
+            ),
+            filter(FilterType::NoEntry, Color::RED, "Modal filter - no entry"),
+            filter(FilterType::BusGate, *colors::BUS_ROUTE, "Bus gate"),
+        ])
+        .section(ctx),
+        ctx.style()
+            .btn_solid_primary
+            .icon("system/assets/tools/select.svg")
+            .disabled(matches!(edit_mode, EditMode::FreehandFilters(_)))
+            .hotkey(Key::F2)
+            .build_widget(ctx, "Freehand filters")
+            .centered_vert(),
+        ctx.style()
+            .btn_solid_primary
+            .icon("system/assets/tools/one_ways.svg")
+            .disabled(matches!(edit_mode, EditMode::Oneways))
+            .hotkey(Key::F3)
+            .build_widget(ctx, "One-ways")
+            .centered_vert(),
+        ctx.style()
+            .btn_solid_primary
+            .icon("system/assets/tools/shortcut.svg")
+            .disabled(matches!(edit_mode, EditMode::Shortcuts(_)))
+            .hotkey(Key::F4)
+            .build_widget(ctx, "Shortcuts")
+            .centered_vert(),
+    ])
 }
